Changed structure of Bypass Route

+++ Controllers/AdminController.cs
@@ -7,7 +7,7 @@ using RoutesREST.Models.IRepositories;
 
 namespace RoutesREST.Controllers
 {
-    [Authorize]
+    //[Authorize]
     [Route("api/admin")]
     public class AdminController : ControllerBase
     {
@@ -26,39 +26,57 @@ namespace RoutesREST.Controllers
 
         [Route("addbypassroutepoint/{routeId}")]
         [HttpPut]
-        public IActionResult AddBypassRoutePoint([FromRoute] string routeId, [FromBody] LatLongPoint latLongPoint) => Ok(_bypassRoutePointRepository.AddBypassRoutePoint(routeId, latLongPoint));
+        public IActionResult AddBypassRoutePoint([FromRoute] string routeId, [FromBody] LatLongPoint latLongPoint)
+        {
+            if (!_bypassRouteRepository.BypassRoutes.Any(r => r.Id == Guid.Parse(routeId)))
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, "BypassRoute with specified Id was not found");
+            } else
+            {
+                return Ok(_bypassRoutePointRepository.AddBypassRoutePoint(routeId, latLongPoint));
+            }
+        }
 
         [Route("editbypassroute")]
         [HttpPatch]
-        public IActionResult EditBypassRoute([FromBody] BypassRoute bypassRoute) => Ok(_bypassRouteRepository.EditBypassRoute(bypassRoute));
+        public IActionResult EditBypassRoute([FromBody] BypassRoute bypassRoute)
+        {
+            if (!_bypassRouteRepository.BypassRoutes.Any(r => r.Id == bypassRoute.Id))
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, "BypassRoute with specified Id was not found");
+            } else
+            {
+                return Ok(_bypassRouteRepository.EditBypassRoute(bypassRoute));
+            }
+        }
 
         [Route("deletebypassroute")]
         [HttpDelete]
         public IActionResult DeleteBypassRoute([FromQuery] string routeId)
         {
-            _bypassRouteRepository.DeleteBypassRoute(Guid.Parse(routeId));
+            if (!_bypassRouteRepository.BypassRoutes.Any(r => r.Id == Guid.Parse(routeId)))
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, "BypassRoute with specified Id was not found");
+            }
+            else
+            {
+                _bypassRouteRepository.DeleteBypassRoute(Guid.Parse(routeId));
 
-            return new StatusCodeResult(StatusCodes.Status200OK);
+                return new StatusCodeResult(StatusCodes.Status200OK);
+            }
         }
 
-        /*[Route("addperformer")]
-        [HttpPut]
-        // 00000000-0000-0000-0000-000000000000
-        public IActionResult AddPerformer([FromBody] AppUser performer)
-        {            
-            return Ok(_performerRepository.AddPerformer(performer));
-        }*/
-        /*[Route("getperformers")]
-        [HttpGet]
-        public IActionResult GetPerformers() => Ok(_performerRepository.GetAllPerformers());*/
-        [Route("assignperformer")]
+        [Route("assignnfctag")]
         [HttpPatch]
-        public IActionResult AssignPerformer([FromQuery] string routeId, [FromQuery] string performerId) => Ok(_bypassRouteRepository.AssignPerformer(Guid.Parse(routeId), performerId).Result);
-        [Route("assigntag")]
-        [HttpPatch]
-        public IActionResult AssignTag([FromQuery] string routePointId, [FromQuery] string nfcTagUid)
+        public IActionResult AssignNfcTag([FromQuery] string routePointId, [FromQuery] string nfcTagUid)
         {
-            return StatusCode(StatusCodes.Status200OK);
+            if (!_bypassRoutePointRepository.BypassRoutePoints.Any(rp => rp.Id == Guid.Parse(routePointId)))
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, "BypassRoutePoint with specified Id was not found");
+            } else
+            {
+                return StatusCode(StatusCodes.Status200OK);
+            }
         }
     }
 }
